+++ assets/javascripts/kitten/components/atoms/status-with-bullet/status-with-bullet.stories.js
@@ -1,12 +1,18 @@
 import React from 'react'
 import { text, select } from '@storybook/addon-knobs'
 import { StatusWithBullet } from './index'
-import { Grid, GridCol } from '../../components/grid/grid'
+import { Grid, GridCol } from '../../../components/grid/grid'
 
 export const Default = () => (
   <StatusWithBullet
     statusMessage={text('Status message', 'Message')}
-    statusType={select('Status type', ['success', 'danger'])}
+    statusType={select(
+      'Status type',
+      ['danger', 'success', 'warning', 'neutral', 'none'],
+      'success',
+    )}
+    size={select('Size', ['normal', 'tiny', 'micro'], 'tiny')}
+    weight={select('Weight', ['light', 'regular', 'bold'], 'regular')}
   />
 )